pybak

+++ pybackup/localcopy.py
@@ -148,15 +148,21 @@ class LocalCopy(OpBase):
                     fdf = tp / rf
                     try:
                         if not tp.exists():
-                            makedirs(tp, exist_ok=True)
-                        fdiff = FileDiff(fsf, fdf)
-                        if fdiff.should_copy():
-                            rv = copy2(di, si, fsf, fdf, self.sfc)
-                            if rv == 0:
-                                print(" ...copied.")
-                                if "exec" in self.opts:
-                                    fdf.chmod(496)
-                                updateDEs(tp, [str(rf)])
+                            rv = makedirs(tp, exist_ok=True)
+                            if rv:
+                                self.sfc.fc += 1
+                        if tp.exists():
+                            fdiff = FileDiff(fsf, fdf)
+                            if fdiff.should_copy():
+                                rv = copy2(di, si, fsf, fdf, self.sfc)
+                                if rv == 0:
+                                    print(" ...copied.")
+                                    self.sfc.sc += 1
+                                    if "exec" in self.opts:
+                                        rv = fdf.chmod(496)
+                                        if rv:
+                                            self.sfc.fc += 1
+                                    updateDEs(tp, [str(rf)])
                     except IOError as exc:
                         print(exc)
                         self.sfc.fc += 1
